Add a Kitchen config file generator
This script looks at a cookbook's metadata's supported platforms and
automatically generates a set of platforms and suites with all the relevant
intermediate_instructions for each platform.

TBD: (How) should this handle cookbooks that, say, only support Systemd
platforms so don't support CentOS 6? That would be addressable but the script
would have to examine supported platforms' versions in the metadata and I don't
really feel like writing code to parse that right this second.