Get rid of coveralls
The gem hasn't been updated in a while and now has a dep conflict on thor with
the chef-dk. This is a stupid reason to get rid of it, but it doesn't exactly
give us much for cookbooks. The Travis build will still fail if the code
coverage isn't sufficient.