Update to the new ChefSpec
We already have one cookbook using the new ChefSpec syntax and will only have
more. The latest releases can no longer converge test cookbooks that use
definitions, but we can hopefully address that on a case-by-case basis by
either updating cookbooks to use custom resources or inserting test stubs.

This can be deleted as soon as a Chef-DK is released packaged with the new
ChefSpec.