Fix Gemfile path finding on Windows
TIL the PATH environment on Windows has semicolons as a separator (presumably
because the paths all have "c:" drive letters in them).

This also removes the chefspec dep, as Chef-DK is now packaged with the version
we need.

It also augments the rubocop dep. Chef-DK now comes with 0.55, which stops
bundler from pulling in one of the newer versions that we've been using.
Further, cookstyle and chefstyle now depend on specific versions of rubocop,
so we can't go to the latest version or those two will blow everything up.